Finalised and tested Term and Descriptor.
Fixed a problem when searching for a '' key in ArangoDB: it raised an error.
Need to decide whether to develop a Node class: the Edge object should contain all that is needed to describe a relationship and may reside in diferent collections, effectively creating non connected graphs. The issue is to what extent should we grant functionality to Tem objects; a category should reference a term or should it be a term?
Need to develop validation procedures based on descriptor data type and kind.