Working on descriptor data validation.
Currently working on getting enumerations connected to descriptors. There is a problem in ArangoDB in the procedure that traverses enumerations connected to descriptors: for some reason it doesn't recognise a term which exists. For that reason there are a couple of echos in the tWrapper.php script.

Need to find a more efficient way of storing enumerations in cached descriptors.

Thought of putting FindByHandle(), CountByHandle() and DeleteByHandle() in the Database class, it is more logical; however there is the formatDocument() call from the document collection that should be either passed to the Database class, or made static.